requirements.txt
# Core dependencies
ell-ai[all]>=0.0.17
python-dotenv>=1.0.0
anthropic>=0.17.0
pydantic>=2.0.0
aiohttp>=3.8.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
isort>=5.0.0
mypy>=1.0.0
flake8>=6.0.0

# Type stubs
types-aiofiles>=0.1.0 